Tests for sumCpp and minor fix.

+++ src/RcppExports.cpp
@@ -6,13 +6,13 @@
 using namespace Rcpp;
 
 // sumCpp
-int sumCpp(int x, int y);
+double sumCpp(double x, double y);
 RcppExport SEXP sumR_sumCpp(SEXP xSEXP, SEXP ySEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
-    Rcpp::traits::input_parameter< int >::type x(xSEXP);
-    Rcpp::traits::input_parameter< int >::type y(ySEXP);
+    Rcpp::traits::input_parameter< double >::type x(xSEXP);
+    Rcpp::traits::input_parameter< double >::type y(ySEXP);
     rcpp_result_gen = Rcpp::wrap(sumCpp(x, y));
     return rcpp_result_gen;
 END_RCPP
